Se craron entidades,repositorio y la opción 2

+++ src/main/java/com/app/monacolibrary/models/DatosLibro.java
@@ -10,6 +10,6 @@ public record DatosLibro(
      @JsonAlias("title") String titulo,
      @JsonAlias("authors") List<DatosAutor> autores,
      @JsonAlias("languages") List<LENGUAJE> idiomas,
-     @JsonAlias("download_count") Integer descargas
+     @JsonAlias("download_count") Long descargas
     ) {
 }
